Set the title of the page to "Don't fall"
Renaming the main module is just bad, and maybe even doesn't work
as it includes characters like ' and " ".

Do as [1] suggests and compile to js instead of html.
This puts the html file under our control, and lets us freely chose
the title. A suggestion from stackoverflow [2] didnt work, maybe as
we target elm 0.17.0 [3].

[1]: https://groups.google.com/forum/#!topic/elm-discuss/4gfEXHDERjM
[2]: http://stackoverflow.com/a/33615549
[3]: https://groups.google.com/forum/#!topic/elm-discuss/khUMddCweEw