New platform generation algorithm
Now we use the murmur hash as primary algorithm for
platform generation. This allows us to not have to
carry a changing seed around. And, more importantly,
we have random access, which allows us to use
List.map instead of recursion (which is not as
beautiful in elm as we can't use pattern matching
like we can in haskel).